Add abstract class PlayerBase
